changes made in dockerfile ad main.py

+++ app_dockerfiles/ubuntu18.04_mysql_server/mysql57_start.sh
@@ -1,10 +1,25 @@
 #!/bin/bash
 
-/etc/init.d/mysql start
+# Start MySQL service
+echo "Starting MySQL service..."
+service mysql start
 
-sleep 10
+# Check if MySQL service started successfully
+if [ $? -ne 0 ]; then
+    echo "Failed to start MySQL service!"
+    exit 1
+fi
 
-mysql --user=root <<EOF
+# Wait for MySQL to be ready
+echo "Waiting for MySQL to start..."
+while ! mysqladmin ping --silent; do
+    sleep 1
+done
+
+echo "MySQL started successfully."
+
+# Run MySQL commands
+mysql --user=root --password=root<<EOF
 ALTER USER 'root'@'localhost' IDENTIFIED BY 'root'; 
 DELETE FROM mysql.user WHERE User='';
 DELETE FROM mysql.user WHERE User='root' AND Host NOT IN ('localhost','127.0.0.1','::1');
@@ -17,6 +32,7 @@ CREATE DATABASE IF NOT EXISTS db_mlflow;
 SHOW DATABASES;
 EOF
 
-  
-tail -f /var/log/mysql/error.log
+echo "All queries runned successfully."
 
+# Tail MySQL error log
+tail -f /var/log/mysql/error.log
